+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # CHANGELOG
 
+## 0.0.16
+
+- FEAT: Callback names are validated.
+
 ## 0.0.15
 Exit and timeout callbacks, output shorthand.
 
 - FEAT: Shorthand `Kommando.puts "uptime"`
 - FEAT: Callback `k.when :exit block`
 - FEAT: Callback `k.when :timeout block`
-- FEAT: Callback names are validated.
 - EXAMPLES: `async.rb`
 
 ## 0.0.14